# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/booking": {
            "get": {
                "description": "Get a list of Booking.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Get All Booking.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Booking"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Booking.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Create New Booking.",
                "parameters": [
                    {
                        "description": "The body to create a new Booking",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.bookingInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Booking"
                        }
                    }
                }
            }
        },
        "/booking/{id}": {
            "get": {
                "description": "Get a Booking by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Get Booking by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Booking"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Booking by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Delete one Booking.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Booking by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Update Booking.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The body to update an Booking",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.bookingInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Booking"
                        }
                    }
                }
            }
        },
        "/booking/{id}/detail": {
            "get": {
                "description": "Get a Booking by ID with Detail.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Get Booking by ID with Detail.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Booking"
                        }
                    }
                }
            }
        },
        "/bookingtype": {
            "get": {
                "description": "Get a list of Booking Type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking Type"
                ],
                "summary": "Get All Booking Type.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BookingType"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Booking Type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking Type"
                ],
                "summary": "Create New Booking Type.",
                "parameters": [
                    {
                        "description": "The body to create a new Booking Type",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.bookingTypeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    }
                }
            }
        },
        "/bookingtype/{id}": {
            "get": {
                "description": "Get a Booking Type by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking Type"
                ],
                "summary": "Get Booking Type by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Booking Type by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking Type"
                ],
                "summary": "Delete one Booking Type.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BookingType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Booking Type by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking Type"
                ],
                "summary": "Update Booking Type.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BookingType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The body to update an Booking Type",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.bookingTypeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BookingType"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.bookingInput": {
            "type": "object",
            "required": [
                "booking_type_id",
                "cars_id",
                "customer_id",
                "end_rent",
                "start_rent"
            ],
            "properties": {
                "booking_type_id": {
                    "type": "integer"
                },
                "cars_id": {
                    "type": "integer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_rent": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "start_rent": {
                    "type": "string"
                }
            }
        },
        "controllers.bookingTypeInput": {
            "type": "object",
            "required": [
                "booking_type",
                "description"
            ],
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
        "models.Booking": {
            "type": "object",
            "required": [
                "booking_type_id",
                "cars_id",
                "customer_id",
                "driver_id",
                "end_rent",
                "start_rent",
                "total_cost"
            ],
            "properties": {
                "booking_type": {
                    "$ref": "#/definitions/models.BookingType"
                },
                "booking_type_id": {
                    "type": "integer"
                },
                "cars": {
                    "$ref": "#/definitions/models.Cars"
                },
                "cars_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/models.Customer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "discount": {
                    "type": "integer"
                },
                "driver": {
                    "$ref": "#/definitions/models.Driver"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_rent": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "start_rent": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "integer"
                },
                "total_driver_cost": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BookingType": {
            "type": "object",
            "required": [
                "booking_type",
                "description"
            ],
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Cars": {
            "type": "object",
            "required": [
                "daily_rent",
                "name",
                "stock"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_rent": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "required": [
                "name",
                "nik",
                "phone_number"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "membership": {
                    "$ref": "#/definitions/models.Membership"
                },
                "membership_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Driver": {
            "type": "object",
            "required": [
                "name",
                "nik",
                "phone_number"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_cost": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Membership": {
            "type": "object",
            "required": [
                "discount",
                "membership_name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "membership_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}